Don't ignore IOErrors on write
Ignoring all IOErrors will mean you don't notice when something didn't work on write. In my case the file was already existing, but the process didn't have write permissions so the changes were not written. If unnoticed you could loose a lot of data with this.

Maybe there is a more sophisticated way to handle this, but until then I think it is better to let all exceptions pass through.