feat(writer): use the mtime option as the date_time value of zip metadata
Before this change, the internal zipfile logic would set date_time in the zip metadata to a timestamp.
This was not a critical problem, but this would mean that generating the same epub file with the same content would lead to different files byte-wise because of different zip metadata.
With this change, if the "mtime" option is set, epub files can now be built in a reproductible way.